requirements.txt
streamlit
torch
torchvision
Pillow
numpy
opencv-python
scipy
matplotlib
pandas
scikit-learn